Sofware Updated Design
This is an assignment-based project. I was given an old design for an application and I had to re-design the same application using FIGMA.
Here in the Folder, I have mentioned the old designs and the updated designs as well. These updated designs are made using modern design rules and laws.